fix manual move

+++ tasks/tasks.h
@@ -1,3 +1,3 @@
 #include <lpc_types.h>
 
-uint8_t _manmove();
+int _manmove();
